fix: implement Petar requests
