+++ __tests__/utils/stark.test.ts
@@ -116,3 +116,12 @@ describe('stark', () => {
     expect(stark.v3Details(detailsUndefined)).toEqual(expect.objectContaining(detailsAnything));
   });
 });
+
+describe('ec full public key', () => {
+  test('determine if value is a BigNumberish', () => {
+    const privateKey1 = '0x43b7240d227aa2fb8434350b3321c40ac1b88c7067982549e7609870621b535';
+    expect(stark.getFullPublicKey(privateKey1)).toBe(
+      '0x0400b730bd22358612b5a67f8ad52ce80f9e8e893639ade263537e6ef35852e5d3057795f6b090f7c6985ee143f798608a53b3659222c06693c630857a10a92acf'
+    );
+  });
+});
