+++ src/utils/stark.ts
@@ -1,6 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import type { SPEC } from 'starknet-types-07';
-import { getStarkKey, utils } from '@scure/starknet';
+import { getPublicKey, getStarkKey, utils } from '@scure/starknet';
 import { gzip, ungzip } from 'pako';
 
 import { ZERO, FeeMarginPercentage } from '../constants';
@@ -14,7 +14,7 @@ import {
 } from '../types';
 import { EDAMode, EDataAvailabilityMode, ETransactionVersion, ResourceBounds } from '../types/api';
 import { FeeEstimate } from '../types/provider';
-import { addHexPrefix, arrayBufferToString, atobUniversal, btoaUniversal } from './encode';
+import { addHexPrefix, arrayBufferToString, atobUniversal, btoaUniversal, buf2hex } from './encode';
 import { parse, stringify } from './json';
 import {
   addPercent,
@@ -350,4 +350,20 @@ export function reduceV2(providedVersion: ETransactionVersion): ETransactionVers
   if (providedVersion === ETransactionVersion.F2) return ETransactionVersion.F1;
   if (providedVersion === ETransactionVersion.V2) return ETransactionVersion.V1;
   return providedVersion;
+}
+
+/**
+ * get the hex string of the full public key related to a Starknet private key.
+ * @param {BigNumberish} privateKey a 252 bits private key.
+ * @returns {string} an hex string of a 520 bit number, representing the full public key related to `privateKey`.
+ * @example
+ * ```typescript
+ * const result = ec.getFullPublicKey("0x43b7240d227aa2fb8434350b3321c40ac1b88c7067982549e7609870621b535");
+ * // result = "0x0400b730bd22358612b5a67f8ad52ce80f9e8e893639ade263537e6ef35852e5d3057795f6b090f7c6985ee143f798608a53b3659222c06693c630857a10a92acf"
+ * ```
+ */
+export function getFullPublicKey(privateKey: BigNumberish): string {
+  const privKey = toHex(privateKey);
+  const fullPrivKey = addHexPrefix(buf2hex(getPublicKey(privKey, false)));
+  return fullPrivKey;
 }
